+++ crates/nix_rs/src/flake/mod.rs
@@ -1,5 +1,6 @@
 //! Rust module for Nix flakes
 
+pub mod command;
 pub mod eval;
 pub mod metadata;
 pub mod outputs;